requirements.txt
dask==2021.9.1
distributed==2021.9.1
pandas==1.3.3
requests==2.22.0
snap-stanford==6.0.0
numpy==1.21.2
seaborn==0.9.0
matplotlib==3.1.0
